requirements.txt
python-dotenv
diskcache
youtube-dl
tqdm
beautifulsoup4
